feat: cursor updating on active tab and some refactor
