--- src/App.module.css
@@ -1,161 +1,0 @@
-.page {
-  min-height: 100vh;
-  background: #e9f3c0;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-orient: vertical;
-  -webkit-box-direction: normal;
-      -ms-flex-direction: column;
-          flex-direction: column;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  padding: 40px 20px;
-  color: white;
-}
-
-.container {
-  background: white;
-  border-radius: 8px;
-  padding: 24px;
-  width: 100%;
-  max-width: 600px;
-  -webkit-box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-          box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-  color: #333;
-}
-
-.inputContainer {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  gap: 8px;
-  margin-bottom: 16px;
-}
-
-.inputContainer input {
-  -webkit-box-flex: 1;
-      -ms-flex: 1;
-          flex: 1;
-  padding: 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-.inputContainer button {
-  padding: 8px 16px;
-  background: #4caf50;
-  color: white;
-  border: none;
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.inputContainer button:hover {
-  background: #45a049;
-}
-
-.todoListContainer {
-  min-height: 100px;
-}
-
-.todoList {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.todoItem {
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  gap: 8px;
-  padding: 8px;
-  border-bottom: 1px solid #eee;
-  background: white;
-}
-
-.todoItem[data-rbd-draggable-context-id] {
-  cursor: -webkit-grab;
-  cursor: grab;
-}
-
-.todoItem[data-rbd-dragging] {
-  background: #f0f0f0;
-  -webkit-box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-          box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-}
-
-.todoItem .rank {
-  min-width: 24px;
-  height: 24px;
-  border-radius: 12px;
-  background: #eee;
-  color: #666;
-  display: -webkit-box;
-  display: -ms-flexbox;
-  display: flex;
-  -webkit-box-align: center;
-      -ms-flex-align: center;
-          align-items: center;
-  -webkit-box-pack: center;
-      -ms-flex-pack: center;
-          justify-content: center;
-  font-size: 12px;
-  cursor: -webkit-grab;
-  cursor: grab;
-}
-
-.todoItem input[type="checkbox"] {
-  margin: 0;
-}
-
-.todoItem input[type="text"] {
-  -webkit-box-flex: 1;
-      -ms-flex: 1;
-          flex: 1;
-  padding: 4px 8px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-.todoItem button {
-  margin-left: auto;
-  background: none;
-  border: none;
-  color: #ff4444;
-  cursor: pointer;
-  font-size: 18px;
-  padding: 4px 8px;
-}
-
-.todoItem button:hover {
-  background: #ffeeee;
-  border-radius: 4px;
-}
-
-.todoItem .editButton {
-  margin-left: 0;
-  color: #666;
-  font-size: 16px;
-}
-
-.todoItem .editButton:hover {
-  background: #eee;
-}
-
-.todoItem.dragging {
-  background: #f0f0f0;
-  -webkit-box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-          box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
-}
-
-.completed {
-  text-decoration: line-through;
-  color: #999;
-}
-/*# sourceMappingURL=App.module.css.map */
