+++ src/styles/app.module.scss
@@ -1,7 +1,7 @@
 .page {
   min-height: 100vh;
   // background: linear-gradient(135deg, #e9f3c0 0%, #6f7246 100%);
-  background: #e9f3c0;
+  background: #ececec;
   display: flex;
   flex-direction: column;
   align-items: center;
@@ -17,32 +17,6 @@
   max-width: 600px;
   box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
   color: #333;
-}
-
-.inputContainer {
-  display: flex;
-  gap: 8px;
-  margin-bottom: 16px;
-
-  input {
-    flex: 1;
-    padding: 8px;
-    border: 1px solid #ddd;
-    border-radius: 4px;
-  }
-
-  button {
-    padding: 8px 16px;
-    background: #4caf50;
-    color: white;
-    border: none;
-    border-radius: 4px;
-    cursor: pointer;
-
-    &:hover {
-      background: #45a049;
-    }
-  }
 }
 
 .todoListContainer {
@@ -100,7 +74,7 @@
     margin-left: auto;
     background: none;
     border: none;
-    color: #ff4444;
+    color: #ec36a6;
     cursor: pointer;
     font-size: 18px;
     padding: 4px 8px;
