+++ src/styles/input.module.scss
@@ -1,0 +1,25 @@
+.inputContainer {
+  display: flex;
+  gap: 8px;
+  margin-bottom: 16px;
+
+  input {
+    flex: 1;
+    padding: 8px;
+    border: 1px solid #ddd;
+    border-radius: 4px;
+  }
+
+  button {
+    padding: 8px 16px;
+    background: #ec36a6;
+    color: white;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+
+    &:hover {
+      background: #be2c86;
+    }
+  }
+}
